Yet an other WBGradient API change. No longer let the user specify layer size. Take only the relevant value (width or height) in user space coords. Using this unique value, the WBGradientBuilder class can create a layer with the best pixel size for the target context (scaling if needed, and forcing non relevant size to 1px).

+++ Sources/Interface/WBHeaderView.m
@@ -165,7 +165,7 @@ static NSImage *kWBHeaderViewGrayBackground = nil;
 }
 
 - (void)wb_addButton:(NSButton *)aButton position:(WBHeaderPosition)position {
-  NSUInteger x = 0;
+  CGFloat x = 0;
   NSUInteger mask = 0;
   if (position == kWBHeaderLeft) {
     mask = NSViewMaxXMargin;
